不知道日版何時變更過的兩個檔案,造成體回無效果?
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@266 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/serverpackets/S_NpcChatPacket.java
@@ -31,11 +31,11 @@ public class S_NpcChatPacket extends ServerBasePacket {
 
 	private byte[] _byte = null;
 
-	public S_NpcChatPacket(L1NpcInstance npc, String chat, byte type) {
+	public S_NpcChatPacket(L1NpcInstance npc, String chat, int type) {
 		buildPacket(npc, chat, type);
 	}
 
-	private void buildPacket(L1NpcInstance npc, String chat, byte type) {
+	private void buildPacket(L1NpcInstance npc, String chat, int type) {
 		switch (type) {
 		case 0: // normal chat
 			writeC(Opcodes.S_OPCODE_NPCSHOUT); // Key is 16 , can use
